fix(ci): rename monogramGlyph to avoid Darwin case clash with RepoMonogram
Case-insensitive FS (and nix-darwin typecheck) treated repoMonogram.ts and
RepoMonogram.tsx as one module, breaking default-export and casing checks.
