+++ packages/client/src/canvas/dock/chipInitials.test.ts
@@ -1,7 +1,7 @@
 import { LOCAL_LOCATION, type TerminalMetadata } from "@kolu/padi/surface";
 import { describe, expect, it } from "vitest";
 import type { TerminalDisplayInfo } from "../../terminal/terminalDisplay";
-import { repoMonogram } from "../../ui/repoMonogram";
+import { repoMonogram } from "../../ui/monogramGlyph";
 import { chipInitials } from "./chipInitials";
 
 function info(group: string, label: string): TerminalDisplayInfo {
